+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Support for multi-drive in CANopen's NetStatusListener.
 - EtherCAT communication via CoE (SDOs).
 - Add image attribute to dictionary class.
+- Add EthercatDictionary class.
+- Add EthercatRegister class.
 - Create EtherCAT PDOs example script.
 
 ### Fixed
@@ -13,6 +15,7 @@
 - Set product name correctly if no dictionary is provided.
 - Docstrings from Register constructor and its subclasses are updated.
 - CanopenRegister and EthernetRegister have the same signature.
+- Add PySOEM to setup.py.
 
 ### Changed
 - Raise ILValueError when the disturbance data does not fit in a register.